requirements.txt
fastapi==0.115.2
uvicorn==0.31.1
SQLAlchemy==2.0.35
pydantic==2.9.2
python-socketio==5.11.4
PyMySQL == 1.1.1
python-dotenv==1.0.1